+++ CreationsForge.Core/Importers/PerkImporter.cs
@@ -26,7 +26,7 @@ public class PerkImporter : ITypedRecordImporter
 
     public string TableName => RecordTypeCatalog.Perk.TableName;
 
-    public IReadOnlySet<SupportedGame> SupportedGames { get; } = new HashSet<SupportedGame> { SupportedGame.Starfield };
+    public IReadOnlySet<SupportedGame> SupportedGames { get; } = new HashSet<SupportedGame> { SupportedGame.Starfield, SupportedGame.Fallout4, SupportedGame.Skyrim };
 
     public void Import(object recordDTO, RecordTypeImportResultDTO result, DateTime importedAtUTC)
     {